Workaround ROOT serialization bug affecting CalDet<PadFlafs>
ROOT has an issue when serializing std::vector<T> where T is a scoped
enum backed by something which has size different from int one.

This is true for any architecture, it just happens to be more lethal on
ARM. For more details https://github.com/root-project/root/issues/16312.

Add explicitly types to the linkdef while at it.

+++ DataFormats/Detectors/TPC/include/DataFormatsTPC/Defs.h
@@ -97,7 +97,7 @@ enum class StatisticsType {
   MeanStdDev   ///< Use mean and standard deviation
 };
 
-enum class PadFlags : unsigned short {
+enum class PadFlags : unsigned int {
   flagGoodPad = 1 << 0,      ///< flag for a good pad binary 0001
   flagDeadPad = 1 << 1,      ///< flag for a dead pad binary 0010
   flagUnknownPad = 1 << 2,   ///< flag for unknown status binary 0100
